pyproject.toml
[project]
name = "readstack-api"
version = "0.1.0"
description = "Readstack API"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.110.0,<1.0.0",
    "sqlalchemy>=2.0.0,<2.1.0",
    "alembic>=1.12.1,<2.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.2.1,<3.0.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "coverage<8.0.0,>=7.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py311"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "B", "C4", "UP", "ARG001",
]
ignore = [
    "E501",
    "B008",
    "W191",
    "B904",
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
